Renaming spec files
